Move parameters to expression
